refactor: seperate reviewList
