+++ frontend/components/rackets/RacketSpec.tsx
@@ -4,7 +4,7 @@ import { Rating } from "react-simple-star-rating";
 import { Tooltip } from "react-tooltip";
 import deepFreeze from "utils/deepFreeze";
 
-interface RacketSpecProps {
+export interface RacketSpecProps {
   score: number;
   weight: Tweight[];
   balance: Tbalance;
@@ -63,15 +63,16 @@ const RacketSpec = ({ score, weight, balance, shaft }: RacketSpecProps) => {
       <div className="flex flex-col justify-center w-mb h-mb gap-y-4">
         <p className="text-2xl font-bold">라켓 스펙</p>
         <Tooltip id="weight">
-          <div style={{ display: "flex", flexDirection: "column" }}>
+          <div className="flex flex-col" data-testid="weight tooltip">
             <span>U 앞에 붙은 숫자가 클수록 가벼운 라켓입니다.</span>
             <span>일반적으로 4U라켓을 사용합니다.</span>
           </div>
         </Tooltip>
-        <div data-tooltip-id="weight">
+        <div data-tooltip-id="weight" data-testid="weight container">
           <p className="text-xl font-bold ">무게</p>
           <div className="flex">
             <p
+              data-testid="W3U"
               className={`flex-1 py-1 text-center ${isWeightSelected(
                 weight,
                 "W3U",
@@ -80,6 +81,7 @@ const RacketSpec = ({ score, weight, balance, shaft }: RacketSpecProps) => {
               3U
             </p>
             <p
+              data-testid="W4U"
               className={`flex-1 py-1 text-center ${isWeightSelected(
                 weight,
                 "W4U",
@@ -88,6 +90,7 @@ const RacketSpec = ({ score, weight, balance, shaft }: RacketSpecProps) => {
               4U
             </p>
             <p
+              data-testid="W5U"
               className={`flex-1 py-1 text-center ${isWeightSelected(
                 weight,
                 "W5U",
@@ -99,23 +102,23 @@ const RacketSpec = ({ score, weight, balance, shaft }: RacketSpecProps) => {
         </div>
 
         <Tooltip id="balance">
-          <div style={{ display: "flex", flexDirection: "column" }}>
+          <div className="flex flex-col" data-testid="balance tooltip">
             <span>밸런스는 라켓의 중심점을 의미합니다.</span>
             <span>
-              {" "}
               동일한 무게라도 밸런스에 따라 다른 느낌을 줄 수 있습니다.
             </span>
-            <ul style={{ listStyle: "inside" }}>
+            <ul className="list-inside list-disc">
               <li>헤드 라이트 : 수비형</li>
               <li>이븐 밸런스 : 올라운드 형</li>
               <li>헤드 헤비 : 공격형</li>
             </ul>
           </div>
         </Tooltip>
-        <div data-tooltip-id="balance">
+        <div data-tooltip-id="balance" data-testid="balance container">
           <p className="text-xl font-bold">밸런스</p>
           <div className="flex">
             <p
+              data-testid="HEAD_LIGHT"
               className={`flex-1 py-1 text-center ${isBalanceSelected(
                 balance,
                 "HEAD_LIGHT",
@@ -124,6 +127,7 @@ const RacketSpec = ({ score, weight, balance, shaft }: RacketSpecProps) => {
               헤드 라이트
             </p>
             <p
+              data-testid="EVEN"
               className={`flex-1 py-1 text-center  ${isBalanceSelected(
                 balance,
                 "EVEN",
@@ -132,6 +136,7 @@ const RacketSpec = ({ score, weight, balance, shaft }: RacketSpecProps) => {
               이븐 밸런스
             </p>
             <p
+              data-testid="HEAD_HEAVY"
               className={`flex-1 py-1 text-center ${isBalanceSelected(
                 balance,
                 "HEAD_HEAVY",
@@ -143,20 +148,20 @@ const RacketSpec = ({ score, weight, balance, shaft }: RacketSpecProps) => {
         </div>
 
         <Tooltip id="flexibility">
-          <div style={{ display: "flex", flexDirection: "column" }}>
+          <div className="flex flex-col" data-testid="flexibility tooltip">
             <span>유연성은 샤프트의 휘어짐의 정도를 의미합니다.</span>
-            <ul style={{ listStyle: "inside" }}>
+            <ul className="list-inside list-disc">
               <li>유연한 라켓일수록 적은 힘으로 멀리 보낼 수 있습니다.</li>
               <li>단단한 라켓일수록 힘을 정확하게 전달할 수 있습니다.</li>
             </ul>
-            <br />
             <span>초보자 분들께는 부드러운 라켓사용을 추천드립니다.</span>
           </div>
         </Tooltip>
-        <div data-tooltip-id="flexibility">
+        <div data-tooltip-id="flexibility" data-testid="flexibility container">
           <p className="text-xl font-bold">탄성</p>
           <div className="flex">
             <p
+              data-testid="FLEXIBLE"
               className={`flex-1 py-1 text-center  ${isShaftSelected(
                 shaft,
                 "FLEXIBLE",
@@ -165,6 +170,7 @@ const RacketSpec = ({ score, weight, balance, shaft }: RacketSpecProps) => {
               부드러움
             </p>
             <p
+              data-testid="MEDIUM"
               className={`flex-1 py-1 text-center ${isShaftSelected(
                 shaft,
                 "MEDIUM",
@@ -173,6 +179,7 @@ const RacketSpec = ({ score, weight, balance, shaft }: RacketSpecProps) => {
               중간
             </p>
             <p
+              data-testid="STIFF"
               className={`flex-1 py-1 text-center ${isShaftSelected(
                 shaft,
                 "STIFF",
