+++ frontend/interface/User.interface.ts
@@ -1,4 +1,4 @@
-interface User {
+export interface User {
   id: number;
   nickname: string;
   email: string;
